requirements.txt
torch
torchvision
datasets
pillow
tkinterdnd2
